Move indexers into an own module
