+++ src/senaite/patient/catalog/configure.zcml
@@ -1,23 +1,7 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    i18n_domain="senaite.patient">
+<configure xmlns="http://namespaces.zope.org/zope"
+           i18n_domain="senaite.patient">
 
-  <!-- Sample (aka AnalysisRequest) Index Adapters -->
-  <adapter name="is_temporary_mrn" factory=".indexers.is_temporary_mrn"/>
-  <adapter name="medical_record_number" factory=".indexers.medical_record_number"/>
-
-  <!-- Patient Indexes -->
-  <adapter name="patient_mrn" factory=".indexers.patient_mrn" />
-  <adapter name="patient_identifier_keys" factory=".indexers.patient_identifier_keys" />
-  <adapter name="patient_identifier_values" factory=".indexers.patient_identifier_values" />
-  <adapter name="patient_race_keys" factory=".indexers.patient_race_keys" />
-  <adapter name="patient_ethnicity_keys" factory=".indexers.patient_ethnicity_keys" />
-  <adapter name="patient_marital_status" factory=".indexers.patient_marital_status" />
-  <adapter name="patient_email" factory=".indexers.patient_email" />
-  <adapter name="patient_email_report" factory=".indexers.patient_email_report" />
-  <adapter name="patient_birthdate" factory=".indexers.patient_birthdate" />
-  <adapter name="patient_fullname" factory=".indexers.patient_fullname" />
-  <adapter name="patient_searchable_text" factory=".indexers.patient_searchable_text" />
-  <adapter name="patient_searchable_mrn" factory=".indexers.patient_searchable_mrn" />
+  <!-- Catalog indexes -->
+  <include package=".indexer" />
 
 </configure>
